+++ R/preprocessing.R
@@ -16,7 +16,7 @@ BinarizeCounts.default <- function(
   verbose = TRUE,
   ...
 ) {
-  if (is(object = object, class2 = 'dgCMatrix')) {
+  if (inherits(x = object, what = 'dgCMatrix')) {
     slot(object = object, name = 'x') <- rep.int(
       x = 1,
       times = length(
@@ -33,6 +33,8 @@ BinarizeCounts.default <- function(
 #' @method BinarizeCounts Assay
 #' @importFrom Seurat GetAssayData SetAssayData
 #' @export
+#' @examples
+#' BinarizeCounts(atac_small[['peaks']])
 BinarizeCounts.Assay <- function(
   object,
   assay = NULL,
@@ -54,6 +56,8 @@ BinarizeCounts.Assay <- function(
 #' @method BinarizeCounts Seurat
 #' @importFrom Seurat GetAssay DefaultAssay
 #' @export
+#' @examples
+#' BinarizeCounts(atac_small)
 BinarizeCounts.Seurat <- function(
   object,
   assay = NULL,
@@ -76,7 +80,9 @@ BinarizeCounts.Seurat <- function(
 #' CreateMotifMatrix
 #'
 #' Create a motif x feature matrix from a set of genomic ranges,
-#' the genome, and a set of position weight matrices
+#' the genome, and a set of position weight matrices.
+#'
+#' Requires that motifmatchr is installed (https://www.bioconductor.org/packages/motifmatchr/).
 #'
 #' @param features A GRanges object containing a set of genomic features
 #' @param pwm A \code{\link[TFBSTools]{PFMatrixList}} or \code{\link[TFBSTools]{PWMatrixList}}
@@ -89,7 +95,6 @@ BinarizeCounts.Seurat <- function(
 #' @param ... Additional arguments passed to \code{\link[motifmatchr]{matchMotifs}}
 #'
 #' @return Returns a sparse matrix
-#' @importFrom motifmatchr matchMotifs motifCounts motifMatches
 #' @export
 #' @examples
 #' \dontrun{
@@ -112,11 +117,20 @@ CreateMotifMatrix <- function(
   sep = c("-", "-"),
   ...
 ) {
-  motif_ix <- matchMotifs(pwms = pwm, subject = features, genome = genome, out = "scores", ...)
+  if (!requireNamespace('motifmatchr', quietly = TRUE)) {
+    stop("Please install motifmatchr. https://www.bioconductor.org/packages/motifmatchr/")
+  }
+  motif_ix <- motifmatchr::matchMotifs(
+    pwms = pwm,
+    subject = features,
+    genome = genome,
+    out = "scores",
+    ...
+  )
   if (use.counts) {
-    motif.matrix <- motifCounts(object = motif_ix)
+    motif.matrix <- motifmatchr::motifCounts(object = motif_ix)
   } else {
-    motif.matrix <- motifMatches(object = motif_ix)
+    motif.matrix <- motifmatchr::motifMatches(object = motif_ix)
     motif.matrix <- as(Class = 'dgCMatrix', object = motif.matrix)
   }
   rownames(motif.matrix) <- GRangesToString(grange = features, sep = sep)
@@ -132,14 +146,11 @@ CreateMotifMatrix <- function(
 #' @param assay Name of assay to use. Default is the active assay.
 #' @param n Number of features to retain (default 20000).
 #' @param verbose Display messages
-#'
 #' @importFrom Seurat DefaultAssay GetAssayData "VariableFeatures<-"
-#'
 #' @return Returns a Seurat object with VariableFeatures set to the randomly sampled features.
+#' @export
 #' @examples
 #' DownsampleFeatures(atac_small, n = 10)
-#'
-#' @export
 DownsampleFeatures <- function(
   object,
   assay = NULL,
@@ -178,8 +189,14 @@ DownsampleFeatures <- function(
 #' @importFrom pbapply pblapply
 #' @importFrom Matrix sparseMatrix
 #' @importFrom Rsamtools TabixFile seqnamesTabix
-#'
 #' @export
+#' @examples
+#' \dontrun{
+#' FeatureMatrix(
+#'   fragments = GetFragments(atac_small),
+#'   features = StringToGranges(rownames(atac_small), sep = c(":", "-"))
+#' )
+#' }
 FeatureMatrix <- function(
   fragments,
   features,
@@ -235,8 +252,18 @@ FeatureMatrix <- function(
   rownames(x = featmat) <- names(x = feature.lookup)
   colnames(x = featmat) <- names(x = cell.lookup)
   if (!is.null(x = cells)) {
-    cells.accept <- intersect(x = cells, y = colnames(x = featmat))
-    return(featmat[, cells.accept])
+    missing.cells <- setdiff(x = cells, y = colnames(x = featmat))
+    if (!(length(x = missing.cells) == 0)) {
+      null.mat <- sparseMatrix(
+        i = c(),
+        j = c(),
+        dims = c(nrow(x = featmat), length(missing.cells))
+      )
+      rownames(x = null.mat) <- rownames(x = featmat)
+      colnames(x = null.mat) <- missing.cells
+      featmat <- cbind(featmat, null.mat)
+    }
+    return(featmat[, cells])
   } else {
     return(featmat)
   }
@@ -260,8 +287,15 @@ FeatureMatrix <- function(
 #'
 #' @importFrom data.table fread fwrite
 #' @importFrom Rsamtools indexTabix bgzip
-#'
 #' @export
+#' @examples
+#' \dontrun{
+#' FilterFragments(
+#'   fragment.path = GetFragments(atac_small),
+#'   cells = colnames(atac_small),
+#'   output.path = "./filtered.tsv"
+#' )
+#' }
 FilterFragments <- function(
   fragment.path,
   cells,
@@ -320,7 +354,7 @@ FilterFragments <- function(
 #' @param assay Name of assay to use
 #' @param min.cutoff Cutoff for feature to be included in the VariableFeatures for the object.
 #' This can be a percentile specified as 'q' followed by the minimum percentile, for example 'q5' to set
-#' the top 95% most common features as the VariableFeatures for the object. Alternatively, this can be
+#' the top 95\% most common features as the VariableFeatures for the object. Alternatively, this can be
 #' an integer specifying the minumum number of cells containing the feature for the feature to be included
 #' in the set of VariableFeatures. For example, setting to 10 will include features in >10 cells in the
 #' set of VariableFeatures. If NULL, include all features in VariableFeatures.
@@ -330,6 +364,8 @@ FilterFragments <- function(
 #' @importFrom stats ecdf
 #' @rdname FindTopFeatures
 #' @export
+#' @examples
+#' FindTopFeatures(object = atac_small[['peaks']][])
 FindTopFeatures.default <- function(
   object,
   assay = NULL,
@@ -352,6 +388,8 @@ FindTopFeatures.default <- function(
 #' @importFrom Seurat GetAssayData VariableFeatures
 #' @export
 #' @method FindTopFeatures Assay
+#' @examples
+#' FindTopFeatures(object = atac_small[['peaks']])
 FindTopFeatures.Assay <- function(
   object,
   assay = NULL,
@@ -417,7 +455,8 @@ FindTopFeatures.Seurat <- function(
 #' @importFrom Seurat GetAssayData AddMetaData
 #'
 #' @export
-#'
+#' @examples
+#' FRiP(object = atac_small, peak.assay = 'peaks', bin.assay = 'bins')
 FRiP <- function(
   object,
   peak.assay,
@@ -462,8 +501,14 @@ FRiP <- function(
 #' @param verbose Display messages
 #'
 #' @importFrom GenomicRanges tileGenome
-#'
 #' @export
+#' @examples
+#' \dontrun{
+#' GenomeBinMatrix(
+#'   fragments = GetFragments(atac_small),
+#'   genome = 'hg19'
+#' )
+#' }
 GenomeBinMatrix <- function(
   fragments,
   genome,
@@ -509,6 +554,10 @@ globalVariables(names = 'cell', package = 'Signac')
 #' @return Returns a Seurat object with added metadata for the ratio of mononucleosomal to nucleosome-free fragments
 #' per cell, and the percentile rank of each ratio.
 #' @export
+#' @examples
+#' \dontrun{
+#' NucleosomeSignal(object = atac_small)
+#' }
 NucleosomeSignal <- function(
   object,
   assay = NULL,
@@ -563,12 +612,20 @@ NucleosomeSignal <- function(
 #' @importFrom BSgenome getSeq
 #' @rdname RegionStats
 #' @export
+#' @examples
+#' \dontrun{
+#' library(BSgenome.Hsapiens.UCSC.hg19)
+#' RegionStats(object = rownames(atac_small), genome = BSgenome.Hsapiens.UCSC.hg19, sep = c(":", "-"))
+#' }
 RegionStats.default <- function(
   object,
   genome,
   verbose = TRUE,
   ...
 ) {
+  if (inherits(x = object, what = 'character')) {
+    object <- StringToGRanges(regions = object, sep = sep)
+  }
   sequence.length <- width(x = object)
   sequences <- getSeq(x = genome, names = object)
   gc <- letterFrequency(x = sequences, letters = 'CG') / sequence.length * 100
@@ -662,10 +719,10 @@ RunTFIDF.default <- function(
   verbose = TRUE,
   ...
 ) {
-  if (is(object = object, class2 = "data.frame")) {
+  if (inherits(x = object, what = "data.frame")) {
     object <- as.matrix(x = object)
   }
-  if (!is(object = object, class2 = "dgCMatrix")) {
+  if (!inherits(x = object, what = "dgCMatrix")) {
     object <- as(object = object, Class = "dgCMatrix")
   }
   if (verbose) {
@@ -696,6 +753,8 @@ RunTFIDF.default <- function(
 #' @rdname RunTFIDF
 #' @method RunTFIDF Assay
 #' @export
+#' @examples
+#' RunTFIDF(atac_small[['peaks']])
 RunTFIDF.Assay <- function(
   object,
   assay = NULL,
@@ -724,6 +783,8 @@ RunTFIDF.Assay <- function(
 #' @rdname RunTFIDF
 #' @method RunTFIDF Seurat
 #' @export
+#' @examples
+#' RunTFIDF(object = atac_small)
 RunTFIDF.Seurat <- function(
   object,
   assay = NULL,
@@ -764,6 +825,20 @@ RunTFIDF.Seurat <- function(
 #'
 #' @return Returns a \code{\link[Seurat]{Seurat}} object
 #' @export
+#' @examples
+#' \dontrun{
+#' library(EnsDb.Hsapiens.v75)
+#' gene.ranges <- genes(EnsDb.Hsapiens.v75)
+#' gene.ranges <- gene.ranges[gene.ranges$gene_biotype == 'protein_coding', ]
+#' tss.ranges <- GRanges(
+#'   seqnames = seqnames(gene.ranges),
+#'   ranges = IRanges(start = start(gene.ranges), width = 2),
+#'   strand = strand(gene.ranges)
+#' )
+#' seqlevelsStyle(tss.ranges) <- 'UCSC'
+#' tss.ranges <- keepStandardChromosomes(tss.ranges, pruning.mode = 'coarse')
+#' TSSEnrichment(object = atac_small, tss.positions = tss.ranges[1:2000])
+#' }
 TSSEnrichment <- function(
   object,
   tss.positions,