+++ NEWS.md
@@ -1,9 +1,25 @@
+# Signac 0.2.0
+
+* First CRAN release
+* Removed dependency on TFBSTools and motifmatchr
+* Renamed `PeriodPlot` to `FragmentHistogram`
+* Removed motif dimension reduction functions
+* Removed motif clustering functions
+* Removed `neighbors` and `reductions` slots from `motif` class
+* Added `motif.names` slot to `motif` class
+* Change DESCRIPTION file so that Bioconductor dependencies are automatically installed
+
+# Signac 0.1.6
+
+* Bug fix for `GetCellsInRegion`
+* Improve documentation
+
 # Signac 0.1.5
 
-* New TSSEnrichment and TSSPlot functions for TSS enrichment scoring
-* New InsertionBias function
-* New options in CoveragePlot for scaling tracks 
-* Major speed improvements for CoveragePlot 
+* New `TSSEnrichment` and `TSSPlot` functions for TSS enrichment scoring
+* New `InsertionBias` function
+* New options in `CoveragePlot` for scaling tracks 
+* Major speed improvements for `CoveragePlot` 
 * Improved documentation (added examples)
 
 # Signac 0.1.4